% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_raster.R
\name{aggregate_raster}
\alias{aggregate_raster}
\title{Compute summary statistics of variables in a data frame for each raster grid cell}
\usage{
aggregate_raster(df, var, fun, grid, out = NULL)
}
\arguments{
\item{df}{Data frame with columns X and Y for UTM coordinates. Coordinate reference system of df should match the coordinate reference system of grid}

\item{var}{Variable(s) in df to summarize}

\item{fun}{A function to compute the summary statistics}

\item{grid}{Either a data frame of raster grid cells with columns of X and Y for UTM coordinates, or a RasterLayer (package raster). Coordinate reference system of df should match the coordinate reference system of grid}

\item{out}{Option to return a SpatRaster (default) or a data frame. For a data.frame, use out='df'.}
}
\value{
SpatRaster or data frame
}
\description{
Note that data points outside of the grid area will be ignored.
}
\examples{
x<-aggregate_raster(dat.rv,c("whake.kg.tow",'redfish.kg.tow'),sum,pgrid4,out='df')
head(x)
x2<-aggregate_raster(dat.rv,c("whake.kg.tow",'redfish.kg.tow'),sum,pgrid4)
x2
}
