% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_points_secr.R
\name{assign_points_secr}
\alias{assign_points_secr}
\title{Assign points to polygon using package secr}
\usage{
assign_points_secr(x, y, polygon, polygon.label)
}
\arguments{
\item{x}{X coordinate (e.g., longitude) with same CRS as polygon}

\item{y}{Y coordinate (e.g., latitude) with same CRS as polygon}

\item{polygon}{Spatvector with geometry polygons}

\item{polygon.label}{The layer containing the name of each polygon in the Spatvector}
}
\value{
Returns a dataframe with nrow = df. There is a column for each polygon with 1 or 0 indicating
whether each point falls within the polygon. There is also a column to indicate which polygon(s) each
point has been assigned to.

Data frame
}
\description{
Before applying function, make sure that the CRS of the polygon matches the CRS of the data points.
Returns a dataframe. There is a column for each polygon with 1 or 0 indicating
whether each point falls within the polygon. There is also a column to indicate which polygon(s) each
point has been assigned to. USE THIS INSTEAD OF \code{assign_points_terra} WHEN THERE ARE OVERLAPPING POLYGONS.
}
\examples{
#poly<-get_shapefile('nafo.clipped')
#df2<-assign_points_secr(dat.ziff[1:50,'longitude'],dat.ziff[1:50,'latitude'],poly,"Label")
#head(df2)
}
\seealso{
assign_points_terra, dat.ziff, get_shapefile
}
