% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_points_terra.R
\name{assign_points_terra}
\alias{assign_points_terra}
\title{Assign points to polygon using package terra}
\usage{
assign_points_terra(x, y, polygon)
}
\arguments{
\item{x}{X coordinate (e.g., longitude) with same CRS as polygon}

\item{y}{Y coordinate (e.g., latitude) with same CRS as polygon}

\item{polygon}{Spatvector with geometry polygons}
}
\value{
Data frame
}
\description{
Before applying function, make sure that the CRS of the polygon matches the CRS of the data points.
Each point must belong to only one polygon, otherwise see assign_points_secr.
Returns the original data frame with a new column of the assigned polygon.
IF THERE ARE OVERLAPPING POLYGONS, USE \code{assign_points_secr} INSTEAD.
}
\examples{
#poly<-get_shapefile('nafo.clipped')
#df2<-assign_points_terra(dat.ziff[1:50,'longitude'],dat.ziff[1:50,'latitude'],poly)
#head(df2)
}
\seealso{
assign_points_secr, dat.ziff, get_shapefile
}
