% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_points_to_nearest_polygon.R
\name{assign_points_to_nearest_polygon}
\alias{assign_points_to_nearest_polygon}
\title{For points that are not within polygons, find the nearest polygon edge}
\usage{
assign_points_to_nearest_polygon(x, y, polygon)
}
\arguments{
\item{x}{X-coordinates (must have same crs as polygon)}

\item{y}{Y-coordinates (must have same crs as polygon)}

\item{polygon}{A spatvector with geometry polygons}
}
\value{
A data frame
}
\description{
For points that are not within polygons, find the nearest polygon edge
}
\examples{
v <- terra::vect(system.file("ex/lux.shp", package="terra"))
pts <- matrix(c(5.812,6.577,5.864,50.126,49.774,49.488), ncol=2)
assign_points_to_nearest_polygon(pts[,1],pts[,2],v)
}
