% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_ziff_dates.R
\name{clean_ziff_dates}
\alias{clean_ziff_dates}
\title{Clean ziff dates}
\usage{
clean_ziff_dates(x)
}
\arguments{
\item{x}{A data frame with columns 'date.landed' and 'date.caught'}
}
\value{
Data frame
}
\description{
Retains original data columns, and adds columns of updated dates.
Step 1) In cases where date.landed was "000000" or was missing, the date.caught is inferred to be the true date landed
Step 2) In remaining cases, if date.landed came earlier than date.caught, the date.caught is inferred to be the true date landed
Step 3) In cases where date.caught was "000000" or missing, the date.landed.updated is inferred to be the true date caught
}
\examples{
clean_ziff_dates(dat.ziff[1:5,])
}
