% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coastline}
\alias{coastline}
\title{Coastline shapefile}
\format{
Retrieved as class SpatVector with get_shapefile('coastline')
}
\usage{
coastline
}
\description{
A shapefile of the coastline, sourced from: https://www.cec.org/north-american-environmental-atlas/political-boundaries-2021/. Accessed with get_shapefile('coastline')
}
\keyword{datasets}
