% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_dms_to_dd.R
\name{convert_dms_to_dd}
\alias{convert_dms_to_dd}
\title{Convert coordinates from degrees minutes seconds to decimal degrees}
\usage{
convert_dms_to_dd(x = NULL, deg = NULL, min = NULL, sec = NULL)
}
\arguments{
\item{x}{Coordinate in format dms. May be numeric or character}

\item{deg}{Degrees. Default is NULL}

\item{min}{Minutes. Default is NULL}

\item{sec}{Seconds. Default is NULL}
}
\description{
Convert coordinates from degrees minutes seconds to decimal degrees
}
\examples{
convert_dms_to_dd(x="64° 10.8’")
convert_dms_to_dd(x="64° 10.8’ 00.00")
convert_dms_to_dd("64°10.8’")
convert_dms_to_dd(deg="64°", min="10.8’", sec="45.6")
convert_dms_to_dd(deg="64°", min="10.8’")
convert_dms_to_dd(deg=64, min=10.8)
convert_dms_to_dd("641045.6")
convert_dms_to_dd(641045.6)
convert_dms_to_dd(deg="64°", min="10’", sec="45.6")
}
