% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat.visreg.fits}
\alias{dat.visreg.fits}
\title{Data frame of visreg fits}
\format{
A data frame with 101 rows and 2 variables.
}
\usage{
dat.visreg.fits
}
\description{
Example data frame
}
\keyword{datasets}
