% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat.ziff}
\alias{dat.ziff}
\title{Example data set: Commercial Landings}
\format{
A data frame with 4559 rows and 16 variables:
\describe{
\item{X}{Easting, UTM NAD83, Zone:20N, Units:km}
\item{Y}{Northing, UTM NAD83, Zone:20N, Units:km}
\item{ancfvn}{Anonymised cfvn}
\item{date.landed}{date.landed}
\item{species.code}{species.code}
\item{round.weight}{round.weight}
\item{main.species.caught}{main.species.caught}
\item{main.species.sought}{main.species.sought}
\item{date.caught}{date.caught}
\item{latitude}{decimal degrees, NAD83, EPSG:4269}
\item{longitude}{decimal degrees, NAD83, EPSG:4269}
\item{gear.class}{gear.class}
\item{year}{year}
\item{year.landed}{year.landed}
\item{month.landed}{month.landed}
\item{day.landed}{day.landed}
}
}
\usage{
dat.ziff
}
\description{
An example data set containing spatial information and species catch weights. Not useful for formal analysis!
}
\keyword{datasets}
