% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_grid_manually.R
\name{format_grid_manually}
\alias{format_grid_manually}
\title{Manually format a prediction grid}
\usage{
format_grid_manually(grid, rep.var = NULL, depth.var = NULL, median.var = NULL)
}
\arguments{
\item{grid}{A prediction grid in which each row is unique}

\item{rep.var}{A variable by which to repeat the rows of the grid, e.g., year}

\item{depth.var}{If the grid needs a column for 'depth_scaled', depth.var is the depths used in the original scaling, e.g., from a data set fitted with a model}

\item{median.var}{If the grid needs a column with the most common level of a predictor, median.var comes from the original data that the model was fitted to}
}
\description{
Returns a data frame
}
\examples{
tmp<-pgrid4[1:3,]
set.seed(678)
mv<-sample(c('a','b'),3,replace=TRUE)
names(sort(table(mv),decreasing=TRUE)[1])
dv<-sample(100:500,50)
format_grid_manually(tmp,rep.var=c(2002,2003),median.var=mv,depth.var=dv)
}
\seealso{
format_grid()
}
