% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_depth.R
\name{get_depth}
\alias{get_depth}
\title{Get depths associated with longitude and latitude}
\usage{
get_depth(lon, lat, crs = "epsg:4269")
}
\arguments{
\item{lon}{Longitude in decimal degrees}

\item{lat}{Latitude in decimal degrees}

\item{crs}{Coordinate reference system for lon and lat. Default is NAD83, EPSG:4269}
}
\description{
Returns depths in meters.
The bathymetry reference used is GEBCO_2024, www.gebco.net.
}
