% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gslSpatial-package.R
\docType{package}
\name{gslSpatial-package}
\alias{gslSpatial}
\alias{gslSpatial-package}
\title{gslSpatial: gslSpatial: An R package to assist with spatial analysis in the Gulf of St. Lawrence region}
\description{
Provides tools to assists with spatial analysis of Research Vessel and Landings data in the Gulf Region. Several functions are tailored for use with package `sdmTMB` (See Anderson et al. (2024) \doi{10.1101/2022.03.24.485545}.) Other functions that do not require `sdmTMB` are useful for more general spatial applications.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JoleneSutton/gslSpatial}
  \item \url{https://jolenesutton.github.io/gslSpatial/}
  \item Report bugs at \url{https://github.com/JoleneSutton/gslSpatial/issues}
}

}
\author{
\strong{Maintainer}: Jolene T. Sutton \email{jolene.sutton@gmail.com}

Authors:
\itemize{
  \item Lysandre Landry
}

}
\keyword{internal}
