% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hex}
\alias{hex}
\title{Hexagonal grid, 10x10km}
\format{
Retrieved as class SpatVector with get_shapefile('hex')
}
\usage{
hex
}
\description{
A shapefile used in the fisheriescape project. CRS = ESRI:102001. Accessed with get_shapefile('hex')
}
\keyword{datasets}
