% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_polygon.R
\name{make_polygon}
\alias{make_polygon}
\title{Make polygons from a data frame of spatial coordinates}
\usage{
make_polygon(df, x, y, grp.cols, crs)
}
\arguments{
\item{df}{A data frame spatial coordinates}

\item{x}{Name of column with x coordinates (e.g., longitude). Must be in order.}

\item{y}{Name of column with y coordinates (e.g., latitude). Must be in order.}

\item{grp.cols}{Index columns for dplyr::group_by. Must index at least one variable, which will be used to identify polygons.}

\item{crs}{The coordinate reference system to assign}
}
\value{
Returns a SpatVector
}
\description{
Make polygons from a data frame of spatial coordinates
}
\examples{
x<-c(-65,-60,-55, -61,-60,-59)
y<-c( 47, 50, 47,  48, 49, 48)
grp<-c(rep('big triangle',3),rep('small triangle',3))
source<-c(1,1,1,3,3,3)
d<-cbind.data.frame(x,y,grp,source)
d
pol<-make_polygon(d,'x','y',c('grp','source'),4326)
terra::plot(pol)
}
