% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mpa}
\alias{mpa}
\title{Oceans Act Marine Protected Areas shapefile}
\format{
Retrieved as class SpatVector with get_shapefile('mpa')
}
\usage{
mpa
}
\description{
A shapefile sourced from https://open.canada.ca, restricted to 3Pn4RSTVn. Accessed with get_shapefile('mpa')
}
\keyword{datasets}
