% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nafo.clipped}
\alias{nafo.clipped}
\title{Northwest Atlantic Fisheries Organization (NFAO) shapefile}
\format{
An object of class \code{character} of length 1.
}
\usage{
nafo.clipped
}
\description{
A shapefile sourced from https://www.nafo.int, condensed to 3Pn4RSTVn. Clipped to land. Accessed with get_shapefile('nafo.clipped')
}
\keyword{datasets}
