% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nafo.unclipped}
\alias{nafo.unclipped}
\title{Northwest Atlantic Fisheries Organization (NFAO) shapefile, not clipped to land}
\format{
Retrieved as class spatVector with get_shapefile('nafo')
}
\usage{
nafo.unclipped
}
\description{
A shapefile sourced from https://www.nafo.int, condensed to 3Pn4RSTVn. Accessed with get_shapefile('nafo')
}
\keyword{datasets}
