% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pgrid4}
\alias{pgrid4}
\title{An example prediction grid for the Gulf of St. Lawrence}
\format{
A data frame with 18885 rows and 7 variables:
\describe{
\item{X}{Easting, UTM NAD83, Zone:20N, Units:km}
\item{Y}{Northing, UTM NAD83, Zone:20N, Units:km}
\item{area}{Area of each grid cell}
\item{longitude}{decimal degrees, NAD83, EPSG:4269}
\item{latitude}{decimal degrees, NAD83, EPSG:4269}
\item{nafo.assigned}{NAFO division or subdivision}
\item{depth}{Depth in meters, based on bathymetry reference GEBCO 2024}
}
}
\usage{
pgrid4
}
\description{
A grid covering 4Pn4RSTVn with 4x4km grid cells.
}
\keyword{datasets}
