% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rv.sgsl}
\alias{rv.sgsl}
\title{Southern Gulf of St. Lawrence September Research Vessel Survey}
\format{
Retrieved as class SpatVector with get_shapefile('rv.sgsl')
}
\usage{
rv.sgsl
}
\description{
A shapefile of the boundaries of the Southern Gulf of St. Lawrence September Research Vessel Survey. Accessed with get_shapefile('rv.sgsl')
}
\keyword{datasets}
