% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rv.winter}
\alias{rv.winter}
\title{Gulf of St. Lawrence Winter Research Vessel Survey}
\format{
Retrieved as class SpatVector with get_shapefile('rv.winter')
}
\usage{
rv.winter
}
\description{
A shapefile of the boundaries of the Gulf of St. Lawrence Winter Research Vessel Survey. Accessed with get_shapefile('rv.winter')
}
\keyword{datasets}
