% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_overlaps.R
\name{sum_overlaps}
\alias{sum_overlaps}
\title{Summarize parts of polygons that overlap one another}
\usage{
sum_overlaps(polys, id, var = NULL)
}
\arguments{
\item{polys}{Spatvector of polygons}

\item{id}{Column that contains polygon names}

\item{var}{Optional numeric variable to sum}
}
\description{
Identifies parts of polygons that overlap, counts the number of layers in overlapping portions, and
optionally sums a user-specified numeric variable in the overlapping portions. Returns a Spatvector in
which new polygons are created from the overlapping portions, and no overlaps remain.
}
\examples{
poly1 <- sf::st_polygon(list(rbind(c(0,-1), c(2,-1), c(2,2), c(0,2), c(0,-1))))
poly2 <- sf::st_polygon(list(rbind(c(1,1), c(3,1), c(3,3), c(1,3), c(1,1))))
poly3 <- sf::st_polygon(list(rbind(c(1.5,0), c(4,0), c(4,2), c(1.5,2), c(1.5,0))))
poly4 <- sf::st_polygon(list(rbind(c(5.5,0), c(8,0), c(8,2), c(5.5,2), c(5.5,0))))
sf_polys <- sf::st_sf(
  id = c('A','B','C','D'),
  value = c(10, 10, 50,8),
  geometry = sf::st_sfc(poly1, poly2, poly3,poly4),
  crs = 4326)
test<-terra::vect(sf_polys)
testit<-sum_overlaps(test,'id')
testit
testit<-sum_overlaps(test,'id','value')
testit
}
