% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_coords.R
\name{transform_coords}
\alias{transform_coords}
\title{Transform (project/change) spatial coordinates from one coordinate reference system to another}
\usage{
transform_coords(lon, lat, from, to)
}
\arguments{
\item{lon}{Longitude}

\item{lat}{Latitude}

\item{from}{Original coordinate reference system}

\item{to}{Desired coordinate reference system}
}
\value{
matrix, array
}
\description{
Transform (project/change) spatial coordinates from one coordinate reference system to another
}
\examples{
lon=c(-55,-57)
lat=c(62,63)
transform_coords(lon,lat,from='epsg:4269',to='ESRI:102001')
}
