% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/union_polygons.R
\name{union_polygons}
\alias{union_polygons}
\title{Union specified polygons within a shapefile.}
\usage{
union_polygons(x, lyr, to.union, new.name = NULL)
}
\arguments{
\item{x}{A shapefile of class SpatVector or a class that can be converted to a SpatVector}

\item{lyr}{Name of layer identifying the polygons to union.}

\item{to.union}{Names of polygons to union.}

\item{new.name}{Which name to assign the unioned polygons. Default is the first item in "to.union".}
}
\description{
Union specified polygons within a shapefile.
}
\examples{
\dontrun{
library(gslSpatial)
library(terra)
librarY(tidyterra)
library(ggplot2)
v <- vect(system.file("ex/lux.shp", package="terra"))
ggplot()+geom_spatvector(data=v)+geom_spatvector_text(data=v,aes(label=NAME_2))
v2<-union_polygons(v,'NAME_2',c('Remich','Luxembourg'),'Remich-Luxembourg')
ggplot()+geom_spatvector(data=v2)+geom_spatvector_text(data=v2,aes(label=NAME_2))
}
}
