% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_dms_to_dd.R
\name{convert_dms_to_dd}
\alias{convert_dms_to_dd}
\title{Convert coordinates from degrees minutes seconds to decimal degrees}
\usage{
convert_dms_to_dd(x = NULL, deg = NULL, min = NULL, sec = NULL)
}
\arguments{
\item{x}{Coordinate in format dms. May be numeric or character}

\item{deg}{Degrees. Default is NULL}

\item{min}{Minutes. Default is NULL}

\item{sec}{Seconds. Default is NULL}
}
\description{
Convert coordinates from degrees minutes seconds to decimal degrees
}
\examples{
convert_dms_to_dd(x=454532.4)
convert_dms_to_dd(x="454532.4")
convert_dms_to_dd(x="45° 45’ 32.4")
convert_dms_to_dd(deg=45, min=45, sec=32.4)
}
