% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nafo.clipped}
\alias{nafo.clipped}
\title{Northwest Atlantic Fisheries Organization (NAFO) shapefile, clipped to land}
\format{
Retrieved as class SpatVector with get_shapefile('nafo.clipped')
}
\usage{
nafo.clipped
}
\description{
A shapefile sourced from https://www.nafo.int, restricted to 3Pn4RSTVn. Accessed with get_shapefile('nafo.clipped')
}
\keyword{datasets}
